refactor: [#27] 입력 UI 분리
1) LottoController에 있는 입력 UI 관련 메소드를 새롭게 생성한 LottoInputDevice로 모두 옮김

2) 추가된 요구사항을 확인하고 수정
